Added all status codes except for Payment and Authentication.

+++ MobRecharge/src/main/java/com/freecharge/controllers/OfferController.java
@@ -1,9 +1,10 @@
 package com.freecharge.controllers;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,24 +24,40 @@ public class OfferController {
 	OfferService offerService;
 	
 	@GetMapping("/")
-	List<Offer> getAllOffers() {
+	ResponseEntity <List<Offer>> getAllOffers() {
 		System.out.println("called");
-		return offerService.getAll();
+		List <Offer> offerList = offerService.getAll();
+		if(offerList.isEmpty())
+			return new ResponseEntity<List<Offer>>(offerList, HttpStatus.NO_CONTENT);
+		else
+			return new ResponseEntity<List<Offer>>(offerList, HttpStatus.FOUND);
 	}
 	
 	@PutMapping("/{oid}")
-	public Optional<Offer> updateOffer(@RequestBody Offer offer, @PathVariable Integer oid) {
-		return offerService.update(offer, oid);
+	public ResponseEntity<String> updateOffer(@RequestBody Offer offer, @PathVariable Integer oid) {
+		if(offerService.isPresent(oid)) {
+			offerService.update(offer, oid);
+			return new ResponseEntity<String>("Updated Successfully", HttpStatus.OK);
+		}
+		return new ResponseEntity<String>("Offer not available", HttpStatus.BAD_REQUEST);
 	}
 	
 	@DeleteMapping("/{oid}")
-	public String deleteOffer(@PathVariable Integer oid) {
-		return offerService.delete(oid);
+	public ResponseEntity<String> deleteOffer(@PathVariable Integer oid) {
+		if(offerService.isPresent(oid)) {
+			offerService.delete(oid);
+			return new ResponseEntity<String>("Deleted Successfully", HttpStatus.OK);
+		}
+		else 
+			return new ResponseEntity<String>("Offer not available", HttpStatus.BAD_REQUEST);
 	}
 	
 	@GetMapping("/{oid}")
-	public Offer getOffer(@PathVariable Integer oid) {
-		return offerService.offer(oid);
+	public ResponseEntity<Offer> getOffer(@PathVariable Integer oid) {
+		if(offerService.isPresent(oid)) 
+			return new ResponseEntity<Offer>(offerService.offer(oid), HttpStatus.OK);
+		else
+			return new ResponseEntity<Offer>((Offer) null, HttpStatus.NOT_FOUND);
 	}
 
 }
